pyproject.toml
[build-system]
build-backend = "uv_build"
requires = ["uv_build>=0.8.3,<0.12.0"]

[dependency-groups]
dev = [
  "cairosvg>=2.7.0",
  "fastembed>=0.4.0",
  "hdbscan>=0.8.0",
  "mypy-extensions>=1.0.0",
  "mypy>=1.0.0",
  "plotly>=5.0.0",
  "pre-commit>=3.0.0",
  "pytest-cov>=4.0.0",
  "pytest>=7.0.0",
  "ruff>=0.1.0",
  "types-setuptools>=80.9.0.20250529",
  "typing-extensions>=4.1.0",
  "umap-learn>=0.5.0"
]

[project]
authors = [{name = "Mike Macpherson"}]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = []
description = "A Python package for working with Lucide icons"
keywords = ["lucide", "icons", "svg", "sqlite"]
license = {text = "MIT"}
name = "python-lucide"
readme = "README.md"
requires-python = ">=3.10"
version = "0.2.25"

[project.optional-dependencies]
search = ["fastembed>=0.4.0"]

[project.scripts]
check-lucide-version = "lucide.dev_utils:print_version_status"
lucide = "lucide.cli:main"
lucide-db = "lucide.cli:main_legacy_db"

[project.urls]
"Bug Tracker" = "https://github.com/mmacpherson/python-lucide/issues"
"Homepage" = "https://github.com/mmacpherson/python-lucide"
"Source Code" = "https://github.com/mmacpherson/python-lucide"

[tool.mypy]
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = false
disallow_untyped_decorators = true
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
python_version = "3.10"
show_column_numbers = true
show_error_context = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
check_untyped_defs = true
disallow_incomplete_defs = false
disallow_untyped_defs = false
module = "tests.*"

[tool.pytest.ini_options]
addopts = "--cov=lucide --cov-report=term-missing --cov-report=xml --cov-report=html"
pythonpath = ["src"]
testpaths = ["tests"]
xfail_strict = true

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.format]
docstring-code-format = true
indent-style = "space"
line-ending = "lf"
quote-style = "double"

[tool.ruff.lint]
ignore = [
  "D203",  # one-blank-line-before-class
  "D213"  # multi-line-summary-second-line
]
select = [
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # isort
  "N",  # pep8-naming
  "UP",  # pyupgrade
  "B",  # flake8-bugbear
  "C4",  # flake8-comprehensions
  "SIM",  # flake8-simplify
  "ARG",  # flake8-unused-arguments
  "ERA",  # eradicate
  "PL",  # pylint
  "RET",  # flake8-return
  "SLF",  # flake8-self
  "RUF",  # ruff-specific rules
  "D"  # pydocstyle
]
unfixable = [
  "F401",  # Don't automatically remove unused imports
  "F841"  # Don't automatically remove unused variables
]

[tool.ruff.lint.isort]
known-first-party = ["lucide"]
section-order = [
  "future",
  "standard-library",
  "third-party",
  "first-party",
  "local-folder"
]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401"]
"scripts/**/*.py" = ["D", "SLF001", "PLR2004", "PLR0915", "PLC0415", "ERA001", "E741", "C408"]
"tests/**/*.py" = ["D", "SLF001", "PLR2004", "ARG005"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.uv.build-backend]
# The module name differs from the normalized package name
module-name = "lucide"
# Include tests and database files in source distribution
source-include = [
  "tests/**",
  "src/lucide/data/lucide-icons.db"
]
